no-issue: db defined reports edit
