+++ packages/desktop/app/views/administration/index.js
@@ -5,4 +5,4 @@ export { PermissionsAdminView } from './PermissionsAdminView';
 export { ReferenceDataAdminView } from './ReferenceDataAdminView';
 export { PatientMergeView } from './patientMerge';
 export { SyncView } from './SyncView';
-export { ReportsAdminView } from './reports';
+export { ReportsAdminView } from './reports/ReportsAdminView';
