--- packages/sync-server/__tests__/admin/data/existing-version-number.json
@@ -1,20 +1,0 @@
-{
-    "versionNumber": 1,
-    "query": "select * from patients limit 0",
-    "queryOptions": {
-      "defaultDateRange": "allTime",
-      "dataSources": [
-        "thisFacility"
-      ],
-      "parameters": [
-        {
-          "parameterField": "ParameterSuggesterSelectField",
-          "label": "Area",
-          "name": "locationGroup",
-          "suggesterEndpoint": "locationGroup"
-        }
-      ]
-    },
-    "status": "published",
-    "notes": "Report doing absolutely nothing"
-  }
