+++ packages/sync-server/config/default.json
@@ -845,9 +845,6 @@
     // NONE is a no-op and does nothing
     "deletionAction": "RENAME"
   },
-  "dbDefinedReports": {
-    "fallbackUser": "admin@tamanu.io"
-  },
   "validateQuestionConfigs": {
     "enabled": false
   },
